Allow user fraction in spawn rate

+++ src/aiolocust/main.py
@@ -16,7 +16,7 @@ def main(
     filename: Annotated[str, typer.Argument(help="The test to run")] = "locustfile.py",
     users: Annotated[int, typer.Option("-u", "--users", help="The number of concurrent VUs")] = 1,
     spawn_rate: Annotated[
-        int | None, typer.Option("-r", "--spawn-rate", help="Rate to spawn users at (users per second).")
+        float | None, typer.Option("-r", "--spawn-rate", help="Rate to spawn users at (users per second).")
     ] = None,
     duration: Annotated[int | None, typer.Option("-d", "--duration", help="Stop the test after X seconds")] = None,
     event_loops: Annotated[
